streamlit_app/requirements.txt
# Streamlit App Requirements
streamlit>=1.28.0
tensorflow>=2.12.0
numpy>=1.21.0
Pillow>=9.0.0
gdown>=4.6.0
requests>=2.28.0